v.1.76. Sound

+++ tools.h
@@ -35,7 +35,7 @@ public:
     static bool copyFile(const QString&, const QString&);
     static bool removeFile(const QString&);
     static void pause(const int, const QString& comment = "");
-    static void sound(const QString&, const float);
+    static void sound(const QString&, const int);
 
 private:
     static QString makeFullPath(const QString&, const QString&);
